Add Fission auth
Add an icon to the navbar to authenticate a user with Fission. When a user
clicks on the icon, they are redirected to the Fission auth lobby where they
can authenticate or not. If they authenticate, their Fission username is sent
over a port into the Elm app and displayed next to the icon.